+++ driver/driver_options_integration_test.go
@@ -21,8 +21,8 @@ func TestServerVersion(t *testing.T) {
 	if version.Distribution == "" {
 		t.Fatalf("expected distribution, got %#v", version)
 	}
-	if !strings.Contains(version.Version, "3.13.0") {
-		t.Fatalf("expected 3.13.0 server, got %#v", version)
+	if !strings.Contains(version.Version, "3.13.6") {
+		t.Fatalf("expected 3.13.6 server, got %#v", version)
 	}
 }